+++ packages/infinity-modules/downtime-log/src/components/toolbar/ToggleSelection.vue
@@ -5,8 +5,8 @@
     v-if="$vuetify.breakpoint.smAndUp"
     :label="
       this.multipleSelection
-        ? 'Disable multiple assignment'
-        : 'Enable multiple assignment'
+        ? $t('downtime.disableAssignment')
+        : $t('downtime.enableAssignment')
     "
   ></v-switch>
 </template>