+++ packages/infinity-modules/user-dashboard/src/components/widgets/oee/PerformanceComparision.vue
@@ -33,7 +33,7 @@ export default {
           type: 'column',
         },
         title: {
-          text: 'Performance comparision',
+          text: this.$t('performanceComparision'),
           style: {
             color: '',
           },